% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_test.R
\name{ksample.test}
\alias{ksample.test}
\title{Automatic test selector for survival data}
\usage{
ksample.test(
  time,
  entry = NULL,
  status = NULL,
  group,
  weights = NULL,
  tests = c("ks", "cvm", "logrank"),
  B = 500,
  plot.curves = FALSE,
  holes = c("holes", "add", "conditional"),
  p = 0,
  q = 0,
  seed = NULL,
  keep.boot = FALSE
)
}
\arguments{
\item{time}{Numeric vector of event or follow-up times.}

\item{entry}{Optional entry time (for left truncation).}

\item{status}{Event indicator (1 = event, 0 = censored).}

\item{group}{Group indicator (factor or vector).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{tests}{Character: "ks", "cvm", "logrank".}

\item{B}{Number of bootstrap replications for approximating the \eqn{p}-value.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the function depicts the functions
to be compared.}

\item{holes}{One of 'holes' (default), 'add', or 'conditional'. See \code{\link[=ltrc.estimator]{ltrc.estimator()}} for details.}

\item{p, q}{Weights for the log-rank-type tests to be performed. If NULL, the classical log-rank.
test is performed.}

\item{seed}{Seed for the bootstrap statistics.}

\item{keep.boot}{Logical, default to FALSE. If TRUE, the statistics computed from the bootstrap resamples
are stored.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov and/or Cramér--von Mises statistics.
\item \code{p.value}: approximated \eqn{p}-value of the tests.
\item \code{stat.boot}: bootstrap statistics for the \eqn{p}-value approximation.
\item \code{logrank}: a matrix with weights p and q, test statistics, and \eqn{p}-values.
\item \code{B}: number of bootstrap replications.
\item \code{tests}: tests performed, as specified in the input.
\item \code{curves}: distribution function estimators of the \eqn{k} samples to be compared.
\item \code{groups}: group indicators.
\item \code{holes}: method for handling holes (if appropriate).
\item \code{stat.boot}: bootstrap statistics (if \code{keep.boot = TRUE}).
}
}
\description{
Performs tests for data subject to left truncation, right censoring or both.
}
\examples{
if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")

aa <- ksample.test(time = channing$age,entry = channing$ageentry,
status = channing$death, group = channing$gender,
weights=rep(1/2,2), B = 200, holes='conditional',
tests=c('ks','cvm','logrank'),plot.curves=TRUE,
p=c(0,1,0.5,0),q=c(0,0,0.5,1),keep.boot = FALSE)

summary(aa)
}

}
