% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_aux_lines.R
\name{lines.estimators}
\alias{lines.estimators}
\title{Add estimators to an existing plot}
\usage{
\method{lines}{estimators}(x, conf.int = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"estimators"}.}

\item{conf.int}{Logical, default to FALSE. If TRUE, pointwise confidence bands
are depicted for the curves.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Adds the survival estimator to an existing plot.
}
