% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_aux_truncation_lines.R
\name{lines.t.estimators}
\alias{lines.t.estimators}
\title{Add estimators to an existing plot}
\usage{
\method{lines}{t.estimators}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"estimators"}.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Adds the survival estimator to an existing plot.
}
