% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_cvm.R
\name{lt.cvm}
\alias{lt.cvm}
\title{Cramér--von Mises-type statistic for the \eqn{k}-sample problem under left truncation}
\usage{
lt.cvm(
  p.sample,
  weights = NULL,
  holes = c("holes", "add", "conditional"),
  plot.curves = FALSE
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, X, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{holes}{Way to handle holes, if present. See \code{\link[=lt.estimator]{lt.estimator()}}
for more details.}

\item{plot.curves}{Logical, default to FALSE. If true, the \eqn{k} distributions to
be compared are depicted.}
}
\value{
Numerical value of the test statistic computed from the \eqn{k} samples.
}
\description{
Computes the Cramér--von Mises statistic adapted to left truncation.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Cramér--von Mises test is defined as
\deqn{D_{CvM} = \sum_{j=1}^k n_j \int \left( F_{jn_j}(t) - F_n(t)\right)^2 dF_n(t),}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input \code{weights}. This function implements the previous test statistic accounting for left
truncation.
}
\examples{
if (requireNamespace("mvna", quietly = TRUE)) {
data("abortion", package = "mvna")
pooled.sample <- abortion[abortion$cause==3, 2:4]
head(pooled.sample)
aa <- lt.cvm(p.sample=pooled.sample,weights=rep(1/2,2),plot.curves=TRUE)
aa
}

}
\references{
Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.
}
