% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_estimator.R
\name{lt.estimator}
\alias{lt.estimator}
\title{Survival function estimator under left truncation}
\usage{
lt.estimator(
  o.sample,
  conf.int = FALSE,
  alpha = 0.05,
  conf.type = c("plain", "loglog"),
  holes = c("holes", "add", "conditional")
)
}
\arguments{
\item{o.sample}{Matrix or data frame with columns (U, X), such that \eqn{U \le X}. If
this input contains more columns, they will not be used.}

\item{conf.int}{Logical, set as default to FALSE. If TRUE, the function
returns a pointwise confidence band.}

\item{alpha}{Pointwise confidence bands with \eqn{1-\alpha} confidence level}

\item{conf.type}{One of 'plain' (default) or 'loglog'. Type of pointwise confidence band.}

\item{holes}{One of 'add', 'conditional', or 'holes' (default) (see details).}
}
\value{
An object with the following content:
\itemize{
\item \code{fail.time}: ordered distinct observed event times
\item \code{estimation}: estimated survival function at each observed event time
\item \code{prob}: jumps of the estimator
\item If \code{conf.int = TRUE}, it also returns the confidence band lower and upper limits
for both the distribution and survival function.
}
}
\description{
Computes the survival function estimator accounting for left truncation.
It also returns pointwise confidence bands and includes the possibility of dealing
with holes in the sample.
}
\details{
The estimator of the survival function when data are subject to left truncation is
defined as (Lynden-Bell, 1971)
\deqn{ S_n(t) = \prod_{x_i \leq t} \left( 1-\frac{d_i}{r_i} \right), }
where \eqn{x_1, \ldots, x_m} are the distinct observed event times,
\eqn{d_i = \# \{j : X_j = x_i\}} and \eqn{r_i = \#\{j : U_j \leq x_i \leq X_j\}}. One
should note that this definition of the risk sets is not equal to that in \code{survival::survfit},
where \eqn{r_i = \#\{j : U_j < x_i \leq X_j\}}, so these two implementations may not be
equal when ties are present in the data.

The variance for the confidence bands follows a Greenwood's-like formula:
\deqn{ Var(S_n(t)) = S_n(t)^2 \sum_{x_i \leq t} \frac{d_i}{n_i(n_i-d_i)}.}
A log-log transformation is also possible. The confidence band is based on the
Gaussianity of the process \eqn{\sqrt{n}(S_n(t) - S(t))}.

The function also includes a way to deal with possible holes. A hole is an event time with associated
unitary risk set, namely \eqn{r_i=1} for some \eqn{ i \in \{1, \ldots, m-1\} }. We exclude the largest
event time, since it is always a hole. If a hole is present
in a sample, the estimator degenerates and takes the value 1 from that time on. Different techniques
are available in the literature for the treatment of holes. For instance, in Stute and Wang (2008),
authors suggest to increase the risk set by 1 (holes='add').
Other references, like Klein and Moeschberger (2003), suggest conditioning on the largest hole different
from the largest event time (holes='conditional'). The possibility of computing the degenerate statistic
is also included (holes='holes').
}
\examples{
if (requireNamespace("mvna", quietly = TRUE)) {
data("abortion", package = "mvna")
pooled.sample <- abortion[abortion$cause==3, 2:4]
plot( lt.estimator(pooled.sample[pooled.sample$group==0,],conf.int=TRUE,conf.type='loglog'),
xlab='Time (in weeks)',ylab='Estimated survival probability' )
lines( lt.estimator(pooled.sample[pooled.sample$group==1,],conf.int=TRUE,conf.type='loglog'),col=2 )
}
}
\references{
Klein, J. P. and Moeschberger, M. L. (2003). Survival Analysis: Techniques for Censored and
Truncated Data. Springer.

Lynden-Bell, D. (1971). A method of allowing for known observational selection in small
samples applied to 3CR quasars. Monthly Notices of the Royal Astronomical Society, 155:95-118.

Stute, W. and Wang, J. L. (2008). The central limit theorem under random truncation. Bernoulli,
14:604–622.
}
