% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_ks.R
\name{lt.ks}
\alias{lt.ks}
\title{Kolmogorov--Smirnov-type statistic for the \eqn{k}-sample problem under left truncation}
\usage{
lt.ks(
  p.sample,
  weights = NULL,
  plot.curves = FALSE,
  holes = c("holes", "add", "conditional")
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, X, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{plot.curves}{Logical, default to FALSE. If true, the \eqn{k} distributions to
be compared are depicted.}

\item{holes}{What to do when a hole is present in a sample. See \code{\link[=lt.estimator]{lt.estimator()}}
for more details.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov statistic.
\item \code{time}: time when the maximum is attained.
}
}
\description{
Computes the Kolmogorov--Smirnov statistic adapted to left truncation.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Kolmogorov--Smirnov test is defined as
\deqn{D_{KS} = \sup_t \sum_{j=1}^k n_j \left( F_{jn_j}(t) - F_n(t)\right)^2,}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for left
truncation. The two-sample version of this statistic was studied in Lago et al. (2025).
Note that, when \eqn{k=2}, one has
\deqn{D_{KS} = (n_1p_2^2 + n_2p_1^2) \left( \sup_t \mid F_{1n_1}(t) - F_{2n_2}(t) \mid \right)^2.}
}
\examples{
if (requireNamespace("mvna", quietly = TRUE)) {
data("abortion", package = "mvna")
pooled.sample <- abortion[abortion$cause==3, 2:4]
head(pooled.sample)
aa <- lt.ks(p.sample=pooled.sample,weights=rep(1/2,2),plot.curves=TRUE)
aa
}


}
\references{
Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.

Lago, A., de Uña-Álvarez, J., & Pardo-Fernández, J. C. (2025). A Kolmogorov--Smirnov-type
test for the two-sample problem with left-truncated data. Test, 34(1), 69-90.
}
