% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_logrank.R
\name{lt.logrank}
\alias{lt.logrank}
\title{Weighted log-rank tests under left truncation}
\usage{
lt.logrank(
  p.sample,
  p = 0,
  q = 0,
  plot.curves = FALSE,
  holes = c("holes", "conditional")
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, X, group).}

\item{p, q}{Nonnegative tuning parameters. If one has length 1 and the other
is a vector, the scalar is recycled.}

\item{plot.curves}{Logical, default to FALSE. If true, the the \eqn{k} distribution functions
to be compared are depicted.}

\item{holes}{How the function deals with holes when estimating the survival function
with the pooled sample. One of 'add' (default), 'conditional', and 'holes'. If no holes
are present in the sample, this argument will not be used.}
}
\value{
A matrix with columns \code{statistic}, detailing the choices of \code{p} and
\code{q}, in addition to the corresponding \code{p.value}.
}
\description{
Computes weighted log-rank-type tests (Fleming–Harrington family)
for \eqn{k} groups with left truncation.
}
\details{
The test statistic is based on the quantity
\deqn{L_j = \sum_{i=1}^m \omega_i \left(d_{ji} - d_i \frac{r_{ji}}{r_i}\right),}
for \eqn{j \in \{1, \ldots, k\}}. Denote by \eqn{x_1, \ldots, x_m} the distinct
observed event times from the \eqn{k} samples. For a time \eqn{x_i} and a group \eqn{j},
\eqn{d_{ji}} denotes the total count of individuals that experience the event,
and \eqn{r_{ji}} is the number of individuals at risk. In addition,
\eqn{d_i = \sum_{j=1}^k d_{ji}} and  \eqn{r_i = \sum_{j=1}^k r_{ji}}, which are
the total number of events observed and the total number of individuals at risk
at time \eqn{x_i}, respectively.

Finally, the weights \eqn{\omega_1, \ldots, \omega_m} are taken as
\deqn{w_i = \hat{S}(x_i)^p (1-\hat{S}(x_i))^q,}
with \eqn{\hat{S}(x_i)} the pooled survival just before the event time. The choice
\eqn{p=0=q} results in the classical version of the log-rank test.

The test statistic is given by
\deqn{(L_1, \ldots, L_{k-1}) \hat{\Sigma}^{-1}(L_1, \ldots, L_{k-1})^\top,}
where \eqn{\hat{\Sigma}} is the matrix with the variance-covariance estimators. Each
term in the matrix includes the correction factor \eqn{(n_j-d_j)/(n_j-1)} for tied
event times. The test statistic follows an asymptotic \eqn{\chi^2_{k-1}} distribution under
the null hypothesis of equal target distributions in the \eqn{k} populations.
}
\examples{
if (requireNamespace("mvna", quietly = TRUE)) {
data("abortion", package = "mvna")
pooled.sample <- abortion[abortion$cause==3, 2:4]
head(pooled.sample)
lt.logrank(p.sample=pooled.sample,p=c(0,1,0.5,5),q=c(0,0,0.5,1),plot.curves=TRUE)
}

}
\references{
Klein, J.P., and Moeschberger, M. L. (2003). Survival Analysis: Techniques for Censored
and Truncated Data. Springer, New York.
}
