% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_cvm_pvalue.R
\name{lt.pv.cvm}
\alias{lt.pv.cvm}
\title{Cramér--von Mises-type test for the \eqn{k}-sample problem under left truncation}
\usage{
lt.pv.cvm(
  p.sample,
  weights = NULL,
  B = 500,
  plot.curves = FALSE,
  holes = c("holes", "add", "conditional"),
  seed = NULL,
  keep.boot = FALSE
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, \eqn{\Delta}, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{B}{Number of bootstrap replications to approximate the \eqn{p}-value.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the \eqn{k} curves to be
compared are depicted.}

\item{holes}{How to handle holes if present in a sample. One of \code{holes} (default),
\code{conditional} or \code{add}. See details in \code{\link[=ltrc.estimator]{ltrc.estimator()}}.}

\item{seed}{Seed for the bootstrap resamples.}

\item{keep.boot}{Logical, default to FALSE. If TRUE, the bootstrap statistics are
returned in the output of the function.}
}
\value{
A list with the following elements:
\itemize{
\item \code{statistic}: numerical value of the Cramér--von Mises test statistic
\item \code{pvalue}: \eqn{p}-value of the test
\item \code{curves}: A list with the estimators of the \eqn{k} distribution functions to be
compared.
\item \code{groups}: Group indicator
\item \code{stat.boot}: If \code{keep.boot = TRUE}, the function returns the statistic evaluated
in the bootstrap resamples.
}
}
\description{
Computes the Cramér--von Mises statistic adapted to left truncation,
and the corresponding \eqn{p}-value.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Cramér--von Mises test is defined as
\deqn{D_{CvM} = \sum_{j=1}^k n_j \int \left( F_{jn_j}(t) - F_n(t)\right)^2 dF_n(t),}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input \code{weights}. This function implements the previous test statistic accounting for left
truncation.

The \eqn{p}-value is approximated by means of the obvious bootstrap accounting for
left truncation, as a particular case of the obvious bootstrap proposed for left-truncated
and right-censored data in Bilker and Wang (1997). The exact
algorithm for the two-sample problem can be consulted in Lago et al. (2025).
}
\examples{
if (requireNamespace("mvna", quietly = TRUE)) {
data("abortion", package = "mvna")
pooled.sample <- abortion[abortion$cause==3, 2:4]
head(pooled.sample)
aa <- lt.pv.cvm(p.sample=pooled.sample,weights=rep(1/2,2),plot.curves=TRUE, B = 200)
aa$pvalue
plot(aa)
}


}
\references{
Bilker, W.B., and Wang, M.C. (1997). Bootstrapping left truncated
and right censored data. Communications in Statistics - Simulation and Computation, 26:141-171.

Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.

Lago, A., de Uña-Álvarez, J., & Pardo-Fernández, J. C. (2025). A Kolmogorov–Smirnov-type test
for the two-sample problem with left-truncated data. TEST, 34, 69-90.
}
