% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_sim.R
\name{lt.sim}
\alias{lt.sim}
\title{Left-truncated data simulation}
\usage{
lt.sim(
  n,
  rtrunc,
  trunc.args = list(),
  rtarget,
  target.args = list(),
  seed = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{rtrunc}{A function that simulates data from a distribution}

\item{trunc.args}{A list with the parameters of the truncation variable}

\item{rtarget}{A function that simulates data from a distribution}

\item{target.args}{A list with the parameters of the target distribution}

\item{seed}{Optional seed.}
}
\value{
A data frame with \eqn{n} rows and two columns: the truncation times are in the first column,
and the event times, in the second one.
}
\description{
Simulates data under left truncation from prespecified variables for the
truncation and target variables.
}
\examples{
aa <- lt.sim(1000,rnorm,trunc.args=list(4,1),rnorm,target.args = list(4,1))
head(aa)
plot(lt.estimator(aa),col=2,xlab='t',ylab='Estimated survival probiblity')
curve(1-pnorm(x,4,1),add=TRUE)

}
