% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_cvm.R
\name{ltrc.cvm}
\alias{ltrc.cvm}
\title{Cramér--von Mises-type statistic for the k-sample problem under left truncation
and right censoring}
\usage{
ltrc.cvm(
  p.sample,
  weights = NULL,
  holes = c("holes", "add", "conditional"),
  plot.curves = FALSE
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, Delta, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{holes}{How to handle holes if present in a sample. One of \code{holes} (default),
\code{conditional} or \code{add}. See
\code{\link[=ltrc.estimator]{ltrc.estimator()}} for details.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the \eqn{k} curves to be
compared are depicted.}
}
\value{
Numerical value of the test statistic computed from the \eqn{k} samples.
}
\description{
Computes the Cramér--von Mises statistic adapted to left truncation and right
censoring.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Cramér--von Mises test is defined as
\deqn{D_{CvM} = \sum_{j=1}^k n_j \int \left( F_{jn_j}(t) - F_n(t)\right)^2 dF_n(t),}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for left
truncation and right censoring, as studied in Lago et al. (2026).
}
\examples{
if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")
data.channing <- data.frame(channing$ageentry, channing$age,
channing$death, channing$gender)
ltrc.cvm(data.channing,weights=rep(1/2,2), holes='conditional', plot.curves = TRUE)
}

}
\references{
Lago, A., Pardo-Fernández, J.C., and de Uña-Álvarez, J. (2026). Kolmogorov--Smirnov
and Cramér--von Mises tests for the k-sample problem for left-truncated and
right-censored data. Lifetime Data Analysis: 32(2).

Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.
}
