% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_estimator.R
\name{ltrc.estimator}
\alias{ltrc.estimator}
\title{Survival function estimator under left truncation and right
censoring}
\usage{
ltrc.estimator(
  o.sample,
  conf.int = FALSE,
  alpha = 0.05,
  conf.type = c("plain", "loglog"),
  holes = c("holes", "add", "conditional")
)
}
\arguments{
\item{o.sample}{Matrix or data frame with columns (U, Y, Delta, group).}

\item{conf.int}{Logical, set as default to FALSE. If set to TRUE, the function
returns a pointwise confidence band.}

\item{alpha}{Pointwise confidence bands with \eqn{1-\alpha} confidence level}

\item{conf.type}{Type of pointwise confidence band. Set as default to \eqn{plain}.}

\item{holes}{One of 'add','conditional', or 'holes' (default) (see details).}
}
\value{
An object with the following content:

fail.time: ordered distinct observed event times

estimation: estimated survival function at each observed event time

prob: jumps of the estimator

If conf.int is set to TRUE, it also returns the confidence band lower and upper limits
}
\description{
Computes the survival function estimator accounting for late entry
times and censored observations. It also returns pointwise confidence bands.
}
\details{
The estimator is defined as
\deqn{ S_n(t) = \prod_{y_i \leq t} \left( 1-\frac{d_i}{n_i} \right), }
where \eqn{y_1, \ldots, y_m} are the distinct observed event times,
\eqn{d_i = \# \{j : Y_j = y_i , \Delta_j=1 \}} and \eqn{n_i = \#\{j : U_j \leq y_i \leq Y_j\}}. It reduces to the
Kaplan--Meier estimator if data are not subject to truncation, and to the Lynden-Bell
estimator if no censoring is present.

The variance for the confidence bands follows a Greenwood's like formula:
\deqn{ Var(S_n(t)) = S_n(t)^2 \sum_{y_i \leq t} \frac{d_i}{n_i(n_i-d_i)}.}
The confidence intervals are based on the asymptotic Gaussianity of the process
\eqn{\sqrt{n} (S_n(t) - S(t))}. Different transformations are also possible.

The function also includes a way to deal with possible holes. A hole is an event time with associated
unitary risk set. If a hole is present in a sample, the estimator degenerates and takes the
value 1 from that time on. Different techniques are available in the literature for the treatment
of holes. For instance, in Stute and Wang (2008), authors suggest to increase the risk set by 1 (holes='add').
Other references, like Klein and Moeschberger (2003), suggest conditioning on the largest hole different
from the largest event time (holes='conditional'). The possibility of computing the degenerate statistic
is also included (holes='holes').
}
\examples{
observed.sample <- ltrc.sim(n=1000,rtrunc=rnorm,rtarget=rnorm,rcens=rexp,
trunc.args=list(4,1),target.args = list(4,1),cens.args = list(1))
est <- ltrc.estimator(observed.sample,conf.int=TRUE,alpha=0.05,conf.type='loglog')
str(est)
plot(est,col='firebrick',xlab='t', ylab='Estimated survival probability')

if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")
o.sample <- channing[channing$gender==1,c(3,4,2)] # subgroup of men
head(o.sample)
plot( ltrc.estimator(o.sample,holes='holes'),col=2,xlab='t',
ylab='Estimated survival probability' ,
main = 'Estimated survival probability for men') # no correction
lines( ltrc.estimator(o.sample,holes='add'),col=3 ) # +1 to risk set
lines( ltrc.estimator(o.sample,holes='conditional'),col=4 )
legend('topright',legend=c('holes="holes" ','holes="add" ',
'holes="conditional"'),col=2:4,lwd=rep(1,3))
}

}
\references{
Klein, J. P. and Moeschberger, M. L. (2003). Survival Analysis: Techniques for Censored and
Truncated Data. Springer

Wang, M.C. (1991). Nonparametric estimation from cross-sectional survival data. Journal of
the American Statistical Association, 86:130-143.

Stute, W. and Wang, J. L. (2008). The central limit theorem under random truncation. Bernoulli,
14:604–622.
}
