% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_ks.R
\name{ltrc.ks}
\alias{ltrc.ks}
\title{Kolmogorov--Smirnov-type statistic for the \eqn{k}-sample problem under left truncation
and right censoring}
\usage{
ltrc.ks(
  p.sample,
  weights = NULL,
  plot.curves = FALSE,
  holes = c("holes", "add", "conditional")
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, Delta, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the \eqn{k} curves to be
compared are depicted.}

\item{holes}{How to handle holes if present in a sample. One of \code{holes} (default),
\code{conditional} or \code{add}. See \code{\link[=ltrc.estimator]{ltrc.estimator()}} for details.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov statistic.
\item \code{time}: time when the maximum is attained.
}
}
\description{
Computes the Kolmogorov--Smirnov statistic adapted to left truncation and right
censoring.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Kolmogorov--Smirnov test is defined as
\deqn{D_{KS} = \sup_t \sum_{j=1}^k n_j \left( F_{jn_j}(t) - F_n(t)\right)^2,}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for left
truncation and right censoring, as studied in Lago et al. (2026).
}
\examples{
if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")

data.channing <- data.frame(channing$ageentry, channing$age,
channing$death, channing$gender)
ltrc.ks(data.channing,weights=rep(1/2,2),holes='conditional')
}

}
\references{
Lago, A., Pardo-Fernández, J.C., and de Uña-Álvarez, J. (2026) Kolmogorov--Smirnov
and Cramér--von Mises tests for the k-sample problem for left-truncated and
right-censored data. Lifetime Data Analysis: 32(2).
}
