% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_ks_pvalue.R
\name{ltrc.pv.ks}
\alias{ltrc.pv.ks}
\title{Kolmogorov--Smirnov-type test for the \eqn{k}-sample problem under left truncation
and right censoring}
\usage{
ltrc.pv.ks(
  p.sample,
  weights = NULL,
  B = 500,
  plot.curves = FALSE,
  holes = c("holes", "add", "conditional"),
  keep.boot = FALSE
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, \eqn{\Delta}, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{B}{Number of bootstrap replications to approximate the \eqn{p}-value.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the \eqn{k} curves to be
compared are depicted.}

\item{holes}{How to handle holes if present in a sample. One of \code{holes} (default),
\code{conditional} or \code{add}. See \code{\link[=ltrc.estimator]{ltrc.estimator()}} for details.}

\item{keep.boot}{Logical, default to FALSE. If TRUE, the function returns the statistics
evaluated in the bootstrap resamples.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov statistic.
\item \code{p.value}: approximated p-value of the tests.
\item \code{curves}: A list with the \eqn{k} estimators of the distribution functions to
be compared.
\item \code{groups}: Group indicator
\item \code{stat.boot}: the statistics computed from the bootstrap resamples. Included in the output
only if \code{keep.boot = TRUE}.
}
}
\description{
Computes the Kolmogorov--Smirnov test adapted to left truncation and right
censoring and the corresponding \eqn{p}-value.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Kolmogorov--Smirnov test is defined as
\deqn{D_{KS} = \sup_t \sum_{j=1}^k n_j \left( F_{jn_j}(t) - F_n(t)\right)^2,}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for left
truncation and right censoring, as studied in Lago et al. (2026).

The \eqn{p}-value is approximated by means of the obvious bootstrap accounting for
left truncation and right censoring, studied in Bilker and Wang (1997). The exact
algorithm for the \eqn{k}-sample problem can be consulted in Lago et al. (2026).
}
\examples{
if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")
data.channing <- data.frame(channing$ageentry,channing$age,
channing$death,channing$gender)
aa <- ltrc.pv.ks(data.channing,weights=rep(1/2,2),
holes = 'conditional', plot.curves = FALSE, B = 200)
plot(aa)
aa$pvalue
}

}
\references{
Bilker, W.B., and Wang, M.C. (1997). Bootstrapping left truncated
and right censored data. Communications in Statistics - Simulation and Computation, 26:141-171.

Lago, A., Pardo-Fernández, J.C., and de Uña-Álvarez, J. (2026). Kolmogorov--Smirnov
and Cramér--von Mises tests for the k-sample problem for left-truncated and
right-censored data. Lifetime Data Analysis: 32(2).

Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.
}
