% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_sim.R
\name{ltrc.sim}
\alias{ltrc.sim}
\title{Simulate left-truncated and right-censored data}
\usage{
ltrc.sim(
  n,
  rtrunc,
  trunc.args = list(),
  rtarget,
  target.args = list(),
  rcens,
  cens.args = list(),
  seed = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{rtrunc}{A function generating truncation times.}

\item{trunc.args}{A list of arguments passed to \code{rtrunc}.}

\item{rtarget}{A function generating target/event times.}

\item{target.args}{A list of arguments passed to \code{rtarget}.}

\item{rcens}{A function generating residual censoring times.}

\item{cens.args}{A list of arguments passed to \code{rcens}.}

\item{seed}{Optional seed.}
}
\value{
A data frame with columns \code{u}, \code{y}, and \code{delta}.
}
\description{
Simulates data under left truncation and right censoring.
}
\examples{
aa <- ltrc.sim(n=500,rtrunc=rnorm,trunc.args=list(4,1),
rtarget=rnorm,target.args = list(4,1),
rcens=rexp,cens.args=list(1),seed=12)
head(aa)

}
