% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_truncation_estimator.R
\name{ltrc.truncation.estimator}
\alias{ltrc.truncation.estimator}
\title{Estimator of the truncation distribution function for data subject to
left truncation and right censoring}
\usage{
ltrc.truncation.estimator(o.sample, holes = c("holes", "add", "conditional"))
}
\arguments{
\item{o.sample}{Matrix or data frame with columns (U, Y, Delta).}

\item{holes}{One of 'add', 'conditional' or 'holes' (default). How to deal with holes
in the estimation of the target variable survival function. See
\code{\link[=ltrc.estimator]{ltrc.estimator()}} for details.}
}
\value{
An object with the observed truncation times, the estimator at those
times, the jumps of the estimator and an estimator of the nontruncation probability
}
\description{
It computes the estimator of the truncation distribution function for data subject to
left truncation and right censoring
}
\details{
The estimator of the distribution function of the truncation variable is given by
\deqn{ G_n(t) = \sum_i \frac{S_n(u_i)^{-1}}{\sum_j S_n(u_j)^{-1} } \mathbb{I}[u_i \leq t],}
where \eqn{u_1, \ldots, u_m} are the distinct truncation times and \eqn{S_n} is the
survival function estimator of the target variable.

The nontruncation probability is defined as
\deqn{ \gamma = \mathbb{P}(U \leq X) = \int G(z)dF(z),}where \eqn{F} is the distribution function of the target
variable. Then, it can be estimated as
\deqn{ \gamma_n = \sum_{i=1}^m \varphi_i G_n(y_i),}where \eqn{y_1, \ldots, y_m} are
the distinct observed event times and \eqn{\varphi_1, \ldots, \varphi_m} are
the weights that the estimator \eqn{S_n} assigns to \eqn{y_1, \ldots, y_m}.

In practice, it is possible to observe truncation times larger than the largest
event time. This would break the formula of the estimator \eqn{G_{n}} defined
above. In such a case, following the recommendations in Wang (1991), we reduce
the distribution function of the truncation variable to a conditional distribution
on the identifiable region, by trimming a small amount of data. This is analogous to
conditioning to times largest than the largest hole when estimating the distribution
function of the target variable.
}
\examples{
aa <- ltrc.sim(n=5000,rtrunc=rnorm,trunc.args = list(4,1),rtarget=rnorm,
target.args=list(4,1),rcens=rexp,cens.args=list(1))
head(aa)
plot(ltrc.truncation.estimator(aa))
curve(pnorm(x,4,1),add=TRUE,col=2)

if (requireNamespace("KMsurv", quietly = TRUE)) {
data("channing", package = "KMsurv")

o.sample <- channing[channing$gender==1,c(3,4,2)] # subgroup of men
plot(ltrc.estimator(o.sample)) # hole in the second event time
plot(ltrc.truncation.estimator(o.sample,holes='add'))
lines(ltrc.truncation.estimator(o.sample,holes='conditional'),col=2)
}

}
\references{
Wang, M.C. (1991). Nonparametric Estimation from Cross-Sectional Survival Data. Journal of
the American Statistical Association, 86:130-143.
}
