% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_aux_plots.R
\name{plot.estimators}
\alias{plot.estimators}
\title{Plot method for estimators objects}
\usage{
\method{plot}{estimators}(x, conf.int = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"estimators"}.}

\item{conf.int}{Logical, default to FALSE. If TRUE, pointwise confidence bands
are depicted for the curves.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a step plot of the survival estimator.
}
