% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_aux_plots_test.R
\name{plot.ksamples_test}
\alias{plot.ksamples_test}
\title{Plot method for k-sample test objects}
\usage{
\method{plot}{ksamples_test}(
  x,
  y = NULL,
  col = NULL,
  lwd = 2,
  lty = 1,
  xlab = "Time",
  ylab = "Survival probability",
  main = "Estimated survival curves",
  legend.pos = "topright",
  ...
)
}
\arguments{
\item{x}{An object of class \code{"ksamples_test"}.}

\item{y}{Ignored.}

\item{col}{Vector of colours for the groups.}

\item{lwd}{Line width.}

\item{lty}{Line type.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{main}{Plot title.}

\item{legend.pos}{Position of the legend.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plots the estimated survival curves for the groups involved in
\code{\link[=ltrc.test]{ltrc.test()}}.
}
