% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltrc_aux_truncation_plots.R
\name{plot.t.estimators}
\alias{plot.t.estimators}
\title{Plot method for estimators objects}
\usage{
\method{plot}{t.estimators}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"estimators"}.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a step plot of the survival estimator.
}
