% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_cvm.R
\name{rc.cvm}
\alias{rc.cvm}
\title{Cramér--von Mises-type statistic for the k-sample problem under
right censoring}
\usage{
rc.cvm(p.sample, weights = NULL, plot.curves = FALSE)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, Delta, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{plot.curves}{Logical, default to FALSE. If TRUE, the \eqn{k} distributions to be compared are depicted}
}
\value{
Numerical value of the test statistic computed from the \eqn{k} samples.
}
\description{
Computes the Cramér--von Mises statistic adapted to left truncation and right
censoring.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Cramér--von Mises test is defined as
\deqn{D_{CvM} = \sum_{j=1}^k n_j \int \left( F_{jn_j}(t) - F_n(t)\right)^2 dF_n(t),}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for
right censoring.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
library(survival)

data <- data.frame(y=stanford2$time,
delta=stanford2$status,
group=as.numeric(stanford2$age<=35)+as.numeric(stanford2$age<=47))
rc.cvm(data,weights=table(data$group)/nrow(data))
}

}
\references{
Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.

Schumacher, M. (1984). Two-Sample Tests of Cramér--von Mises-and Kolmogorov--Smirnov-Type
for Randomly Censored Data. International Statistical Review/Revue Internationale de
Statistique, 263-281.
}
