% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_estimator.R
\name{rc.estimator}
\alias{rc.estimator}
\title{Survival function estimator under left truncation and right
censoring}
\usage{
rc.estimator(
  o.sample,
  conf.int = FALSE,
  alpha = 0.05,
  conf.type = c("plain", "loglog")
)
}
\arguments{
\item{o.sample}{Matrix or data frame with columns (U, Y, Delta, group).}

\item{conf.int}{Logical, set as default to FALSE. If set to TRUE, the function
returns a pointwise confidence band.}

\item{alpha}{Pointwise confidence bands with \eqn{1-\alpha} confidence level}

\item{conf.type}{Type of pointwise confidence band. Set as default to \eqn{plain}.}
}
\value{
An object with the following content:

fail.time: ordered distinct observed event times

estimation: estimated survival function at each observed event time

prob: jumps of the estimator

If conf.int is set to TRUE, it also returns the confidence band lower and upper limits
}
\description{
Computes the survival function estimator accounting for censored observations.
It also returns pointwise confidence bands.
}
\details{
The estimator is defined as
\deqn{ S_n(t) = \prod_{y_i \leq t} \left( 1-\frac{d_i}{n_i} \right), }
where \eqn{y_1, \ldots, y_m} are the distinct observed event times,
\eqn{d_i = \# \{j : Y_j = y_i , \Delta_j=1 \}} and \eqn{n_i = \#\{j : U_j \leq y_i \leq Y_j\}}. It reduces to the
Kaplan--Meier estimator if data are not subject to truncation, and to the Lynden-Bell
estimator if no censoring is present.

The variance for the confidence bands follows a Greenwood's like formula:
\deqn{ Var(S_n(t)) = S_n(t)^2 \sum_{y_i \leq t} \frac{d_i}{n_i(n_i-d_i)}.}
Different transformations are also possible.
}
\examples{
x <- rnorm(500,4,1)
c <- rnorm(500,4,1)
o.sample <- data.frame(y = pmin(x,c),delta = as.numeric(x <= c))
plot(rc.estimator(o.sample,conf.int=TRUE,conf.type='loglog'))
curve(1-pnorm(x,4,1),add=TRUE,col=2)

}
\references{
Klein, J. P. and Moeschberger, M. L. (2003). Survival Analysis: Techniques for Censored and
Truncated Data. Springer

Wang, M.C. (1991). Nonparametric estimation from cross-sectional survival data. Journal of
the American Statistical Association, 86:130-143.

Stute, W. and Wang, J. L. (2008). The central limit theorem under random truncation. Bernoulli,
14:604–622.
}
