% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_ks.R
\name{rc.ks}
\alias{rc.ks}
\title{Kolmogorov--Smirnov-type statistic for the k-sample problem under right censoring}
\usage{
rc.ks(p.sample, weights = NULL, plot.curves = FALSE)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (Y, Delta, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{plot.curves}{Logical, FALSE by default. Otherwise, it represents the k distribution
functions to be compared.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov statistics.
\item \code{time}: time when the maximum is attained.
}
}
\description{
Computes the Kolmogorov--Smirnov statistic adapted to right censoring.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Kolmogorov--Smirnov test is defined as
\deqn{D_{KS} = \sup_t \sum_{j=1}^k n_j \left( F_{jn_j}(t) - F_n(t)\right)^2,}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for
right censoring.

For the particular case of \eqn{k=2}, the test statistic reduces to
\deqn{D_{KS} = (n_1p_2^2 + n_2p_1^2) \left(\sup_t \mid F_{1n_1}(t) - F_{2n_2}(t) \mid\right)^2,}
thus the test indeed extends the classical two-sample Kolmogorov--Smirnov test. A two-sample
Kolmogorov--Smirnov-type test for right-censored data was proposed in Schumacher (1984).
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
library(survival)

data <- data.frame(y=stanford2$time,
delta=stanford2$status,
group=as.numeric(stanford2$age<=35)+as.numeric(stanford2$age<=47))
rc.ks(p.sample=data,weights=table(data$group)/nrow(data),plot.curves=TRUE)
}

}
\references{
Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.

Schumacher, M. (1984). Two-Sample Tests of Cramér--von Mises-and Kolmogorov--Smirnov-Type
for Randomly Censored Data. International Statistical Review/Revue Internationale de
Statistique, 263-281.
}
