% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_cvm_pvalue.R
\name{rc.pv.cvm}
\alias{rc.pv.cvm}
\title{Cramér--von Mises-type test for the \eqn{k}-sample problem under right censoring}
\usage{
rc.pv.cvm(p.sample, weights = NULL, B = 500, keep.boot = FALSE)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, \eqn{\Delta}, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{B}{Number of bootstrap replications to approximate the \eqn{p}-value.}

\item{keep.boot}{Logical, default to FALSE. If TRUE, the function returns the evaluation
of the Cramér--von Mises statistic on the bootstrap resamples.}
}
\value{
A list with the following:
\itemize{
\item \code{statistic}: Cramér--von Mises statistic evaluated on the the samples.
\item \code{pvalue}: \eqn{p}-value of the test approximated via bootstrap.
\item \code{curves}: A list with the estimators of the \eqn{k} distribution functions to
be compared.
\item \code{groups}: Group indicator.
\item \code{stat.boot}: The Cramér--von Mises statistic evaluated from the bootstrap resamples.
}
}
\description{
Computes the Kolmogorov--Smirnov test adapted to left truncation and right
censoring and the corresponding \eqn{p}-value.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Cramér--von Mises test is defined as
\deqn{D_{CvM} = \sum_{j=1}^k n_j \int \left( F_{jn_j}(t) - F_n(t)\right)^2 dF_n(t),}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for
right censoring.

The \eqn{p}-value is approximated by means of the obvious bootstrap accounting for
left truncation and right censoring, studied in Efron (1981).
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
library(survival)
data <- data.frame(y=stanford2$time,
delta=stanford2$status,
group=as.numeric(stanford2$age<=35)+as.numeric(stanford2$age<=47))
aa <- rc.pv.cvm(data,weights=table(data$group)/nrow(data), B = 200)
plot(aa)
aa$pvalue
}

}
\references{
Efron, B. (1981). Censored data and the bootstrap. Journal of the American
Statistical Association, 76: 312-319.

Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.
}
