% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_ks_pvalue.R
\name{rc.pv.ks}
\alias{rc.pv.ks}
\title{Kolmogorov--Smirnov-type test for the \eqn{k}-sample problem under right censoring}
\usage{
rc.pv.ks(
  p.sample,
  weights = NULL,
  B = 500,
  plot.curves = FALSE,
  keep.boot = FALSE
)
}
\arguments{
\item{p.sample}{Matrix or data frame with columns (U, Y, \eqn{\Delta}, group).}

\item{weights}{Vector of weights of the estimator under the null hypothesis.
If no vector is provided, weights proportional to each sample size are assigned automatically
by the function. Not used when \code{tests="logrank"}.}

\item{B}{Number of bootstrap replications to approximate the \eqn{p}-value}

\item{plot.curves}{logical, set to FALSE by default. If TRUE, the function depicts
the \eqn{k} functions to be compared.}

\item{keep.boot}{Logical, default to FALSE. If TRUE, the statistics evaluated on the
bootstrap resamples are included in the output.}
}
\value{
A list with the following components:
\itemize{
\item \code{statistic}: numerical value of the Kolmogorov--Smirnov statistic.
\item \code{pvalue}: approximated p-value of the test.
\item \code{curves}: A list with the estimators of the \eqn{k} distribution functions to
be compared.
\item \code{groups}: Group indicator
\item \code{stat.boot}: statistics computed from the bootstrap resamples.
}
}
\description{
Computes the Kolmogorov--Smirnov test adapted to right censoring and the
corresponding \eqn{p}-value.
}
\details{
In Kiefer (1959), the \eqn{k}-sample version of the Kolmogorov--Smirnov test is defined as
\deqn{D_{KS} = \sup_t \sum_{j=1}^k n_j \left( F_{jn_j}(t) - F_n(t)\right)^2,}
where \eqn{F_{jn_j}} is the cumulative distribution function estimator of calculated
only with the \eqn{j}-th sample, and \eqn{F_n} is given by
\deqn{F_n(t) =\sum_{j=1}^k p_j F_{jn_j}(t),}where \eqn{p_1, \ldots, p_k} correspond to the
input weights. This function implements the previous test statistic accounting for
right censoring.

For the particular case of \eqn{k=2}, the test statistic reduces to
\deqn{D_{KS} = (n_1p_2^2 + n_2p_1^2) \left(\sup_t \mid F_{1n_1}(t) - F_{2n_2}(t) \mid\right)^2,}
thus the test indeed extends the classical two-sample Kolmogorov--Smirnov test. A two-sample
Kolmogorov--Smirnov-type test for right-censored data was proposed in Schumacher (1984).

The \eqn{p}-value is approximated by means of the obvious bootstrap accounting for
right censoring, studied in Efron (1981).
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
library(survival)

data <- data.frame(y=stanford2$time,
delta=stanford2$status,
group=as.numeric(stanford2$age<=35)+as.numeric(stanford2$age<=47))
aa <- rc.pv.ks(data,weights=table(data$group)/nrow(data), B = 200)
plot(aa)
aa$pvalue
}

}
\references{
Efron, B. (1981). Censored data and the bootstrap. Journal of the American
Statistical Association, 76: 312-319.

Kiefer, J. (1959). K-sample analogues of the Kolmogorov--Smirnov and Cramér--von Mises
tests. The Annals of Mathematical Statistics, 30:420-447.
}
