% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_sim.R
\name{rc.sim}
\alias{rc.sim}
\title{Right-censored data simulation}
\usage{
rc.sim(
  n,
  rtarget,
  target.args = list(),
  rcens,
  cens.args = list(),
  seed = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{rtarget}{A function that simulates event/survival times.}

\item{target.args}{A list with the parameters of the target distribution.}

\item{rcens}{A function that simulates censoring times.}

\item{cens.args}{A list with the parameters of the censoring distribution.}

\item{seed}{Optional seed for reproducibility.}
}
\value{
A data frame with \eqn{n} rows and two columns:
\describe{
\item{y}{Observed follow-up time, i.e. \eqn{\min(X, C)}.}
\item{delta}{Event indicator, equal to 1 if the event is observed and 0 if censored.}
}
}
\description{
Simulates right-censored survival data from prespecified event-time and
censoring-time distributions.
}
\examples{
aa <- rc.sim(n = 1000, rtarget = rweibull,
target.args = list(shape = 2, scale = 3),
rcens = rexp, cens.args = list(rate = 0.2),
seed = 123)
head(aa)
mean(aa$delta == 0)  # empirical censoring rate

}
